changed srt to use floats to compare

+++ src/process/include/pid_record.h
@@ -19,7 +19,7 @@ typedef struct pid_record_t {
     uint32_t running_time_until_first_response;
     uint32_t first_response_time;
     uint32_t added_to_queue;
-    uint32_t* exp_time_remaining_chart;
+    float* exp_time_remaining_chart;
     uint32_t last_preempted;
     uint32_t* current_time;
